refactor string types - Split the serializers out to /serializers - Left the dispatchers in complex_types (to be moved soon)
