+++ structured/serializers/__init__.py
@@ -1,4 +1,5 @@
 from .api import *
+from .strings import *
 from .structs import *
 from .structured import *
 from .unions import *
